Refactor loadTasks() function
